// angular/BasicDemo/src/app/interfaces/UILayerTypes.ts
export enum UILayerTypes {
    WMSLayer = "WMSLayer",
    PanoramicLayer = "PanoramicLayer",
    HSPCLayer = "HSPCLayer",
    OGC3DTILES = "OGC3DTILES",
    LTSLayer = "LTSLayer",
    TMSLayer = "TMSLayer",
    WFSLayer = "WFSLayer",
    BingmapsLayer="BingmapsLayer",
    LayerGroup="LayerGroup",
    GeoJSONLayer="GeoJSONLayer",
    CustomPanoramicLayer="CustomPanoramicLayer",
}